get puul

+++ src/main/java/utils/PathParser.java
@@ -3,8 +3,8 @@ package utils;
 public class PathParser {
     private PathParser() {}
 
-    public static int getPathVariable(String path) {
+    public static Long getPathVariable(String path) {
         String firstLine = path.split("\n")[0].replaceAll("/", "");
-        return Integer.parseInt(firstLine);
+        return Long.parseLong(firstLine);
     }
 }
